+++ src/onegov.user/onegov/user/models/user.py
@@ -18,6 +18,16 @@ class User(Base, TimestampMixin, ORMSearchable):
 
     __tablename__ = 'users'
 
+    #: the type of the item, this can be used to create custom polymorphic
+    #: subclasses of this class. See
+    #: `<http://docs.sqlalchemy.org/en/improve_toc/\
+    #: orm/extensions/declarative/inheritance.html>`_.
+    type = Column(Text, nullable=True)
+
+    __mapper_args__ = {
+        'polymorphic_on': type
+    }
+
     es_language = 'de'  # XXX add to database in the future
     es_properties = {
         'username': {'type': 'text'},